+++ relay-server/src/services/store.rs
@@ -33,8 +33,8 @@ use relay_threading::AsyncPool;
 use sentry_protos::snuba::v1::any_value::Value;
 use sentry_protos::snuba::v1::{AnyValue, TraceItem, TraceItemType};
 use serde::{Deserialize, Serialize};
-use serde_json::Deserializer;
 use serde_json::value::RawValue;
+use serde_json::{Deserializer, Value as JsonValue};
 use uuid::Uuid;
 
 use crate::metrics::{ArrayEncoding, BucketEncoder, MetricOutcomes};
@@ -893,6 +893,7 @@ impl StoreService {
         item: &Item,
     ) -> Result<(), StoreError> {
         relay_log::trace!("Producing span");
+
         let payload = item.payload();
         let d = &mut Deserializer::from_slice(&payload);
         let mut span: SpanKafkaMessage = match serde_path_to_error::deserialize(d) {
@@ -928,6 +929,27 @@ impl StoreService {
                 .retain(|_, v| v.as_ref().and_then(|v| v.value).is_some_and(f64::is_finite));
         }
 
+        self.inner_produce_item_span(scoping, received_at, event_id, retention_days, span.clone())?;
+        self.inner_produce_span(scoping, span)?;
+
+        self.outcome_aggregator.send(TrackOutcome {
+            category: DataCategory::SpanIndexed,
+            event_id: None,
+            outcome: Outcome::Accepted,
+            quantity: 1,
+            remote_addr: None,
+            scoping,
+            timestamp: received_at,
+        });
+
+        Ok(())
+    }
+
+    fn inner_produce_span(
+        &self,
+        scoping: Scoping,
+        span: SpanKafkaMessage,
+    ) -> Result<(), StoreError> {
         self.produce(
             KafkaTopic::Spans,
             KafkaMessage::Span {
@@ -939,15 +961,267 @@ impl StoreService {
             },
         )?;
 
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::SpanIndexed,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: 1,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
+        Ok(())
+    }
+
+    fn inner_produce_item_span(
+        &self,
+        scoping: Scoping,
+        received_at: DateTime<Utc>,
+        event_id: Option<EventId>,
+        retention_days: u16,
+        span: SpanKafkaMessage,
+    ) -> Result<(), StoreError> {
+        let mut trace_item = TraceItem {
+            item_type: TraceItemType::Span.into(),
+            organization_id: scoping.organization_id.value(),
+            project_id: scoping.project_id.value(),
+            received: Some(Timestamp {
+                seconds: safe_timestamp(received_at) as i64,
+                nanos: 0,
+            }),
+            retention_days: retention_days.into(),
+            timestamp: Some(Timestamp {
+                seconds: span.start_timestamp_precise as i64,
+                nanos: 0,
+            }),
+            trace_id: span.trace_id.to_string(),
+            item_id: u128::from_str_radix(span.span_id, 16)
+                .unwrap_or_default()
+                .to_le_bytes()
+                .to_vec(),
+            attributes: Default::default(),
+            client_sample_rate: 1.0,
+            server_sample_rate: 1.0,
+        };
+
+        if let Some(data) = span.data {
+            for (key, value) in data {
+                let any_value = match value {
+                    JsonValue::String(string) => AnyValue {
+                        value: Some(Value::StringValue(string)),
+                    },
+                    JsonValue::Number(number) => {
+                        if number.is_i64() || number.is_u64() {
+                            AnyValue {
+                                value: Some(Value::IntValue(number.as_i64().unwrap_or_default())),
+                            }
+                        } else {
+                            AnyValue {
+                                value: Some(Value::DoubleValue(
+                                    number.as_f64().unwrap_or_default(),
+                                )),
+                            }
+                        }
+                    }
+                    JsonValue::Bool(bool) => AnyValue {
+                        value: Some(Value::BoolValue(bool)),
+                    },
+                    JsonValue::Array(value) => AnyValue {
+                        value: Some(Value::StringValue(
+                            serde_json::to_string(&value).unwrap_or_default(),
+                        )),
+                    },
+                    JsonValue::Object(value) => AnyValue {
+                        value: Some(Value::StringValue(
+                            serde_json::to_string(&value).unwrap_or_default(),
+                        )),
+                    },
+                    _ => continue,
+                };
+
+                trace_item.attributes.insert(key.into(), any_value);
+            }
+        }
+
+        if let Some(measurements) = span.measurements {
+            for (key, measurement) in measurements {
+                let Some(m) = measurement else {
+                    continue;
+                };
+                let Some(value) = m.value else {
+                    continue;
+                };
+                match key {
+                    Cow::Borrowed("client_sample_rate") if value > 0.0 => {
+                        trace_item.client_sample_rate = value
+                    }
+                    Cow::Borrowed("server_sample_rate") if value > 0.0 => {
+                        trace_item.server_sample_rate = value
+                    }
+                    _ => {
+                        trace_item.attributes.insert(
+                            key.into(),
+                            AnyValue {
+                                value: Some(Value::DoubleValue(value)),
+                            },
+                        );
+                    }
+                }
+            }
+        }
+
+        if let Some(sentry_tags) = span.sentry_tags {
+            for (name, tag) in sentry_tags {
+                let Some(value) = tag else {
+                    continue;
+                };
+
+                let key = if name == "description" {
+                    "sentry.normalized_description".to_string()
+                } else {
+                    format!("sentry.{name}")
+                };
+
+                trace_item.attributes.insert(
+                    key.to_owned(),
+                    AnyValue {
+                        value: Some(Value::StringValue(value)),
+                    },
+                );
+            }
+        }
+
+        if let Some(tags) = span.tags {
+            for (key, tag) in tags {
+                if let Some(value) = tag {
+                    trace_item.attributes.insert(
+                        key.into(),
+                        AnyValue {
+                            value: Some(Value::StringValue(value)),
+                        },
+                    );
+                }
+            }
+        }
+
+        if let Some(description) = span.description {
+            trace_item.attributes.insert(
+                "sentry.raw_description".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(description)),
+                },
+            );
+        }
+
+        trace_item.attributes.insert(
+            "sentry.duration_ms".into(),
+            AnyValue {
+                value: Some(Value::IntValue(span.duration_ms.into())),
+            },
+        );
+
+        if let Some(event_id) = event_id {
+            trace_item.attributes.insert(
+                "sentry.event_id".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(event_id.0.as_simple().to_string())),
+                },
+            );
+        }
+
+        trace_item.attributes.insert(
+            "sentry.is_segment".into(),
+            AnyValue {
+                value: Some(Value::BoolValue(span.is_segment)),
+            },
+        );
+
+        trace_item.attributes.insert(
+            "sentry.exclusive_time_ms".into(),
+            AnyValue {
+                value: Some(Value::DoubleValue(span.exclusive_time_ms)),
+            },
+        );
+
+        trace_item.attributes.insert(
+            "sentry.start_timestamp_precise".into(),
+            AnyValue {
+                value: Some(Value::DoubleValue(span.start_timestamp_precise)),
+            },
+        );
+
+        trace_item.attributes.insert(
+            "sentry.end_timestamp_precise".into(),
+            AnyValue {
+                value: Some(Value::DoubleValue(span.end_timestamp_precise)),
+            },
+        );
+
+        trace_item.attributes.insert(
+            "sentry.start_timestamp_ms".into(),
+            AnyValue {
+                value: Some(Value::IntValue(span.start_timestamp_ms as i64)),
+            },
+        );
+
+        trace_item.attributes.insert(
+            "sentry.is_remote".into(),
+            AnyValue {
+                value: Some(Value::BoolValue(span.is_remote)),
+            },
+        );
+
+        if let Some(parent_span_id) = span.parent_span_id {
+            trace_item.attributes.insert(
+                "sentry.parent_span_id".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(parent_span_id.to_owned())),
+                },
+            );
+        }
+
+        if let Some(profile_id) = span.profile_id {
+            trace_item.attributes.insert(
+                "sentry.profile_id".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(profile_id.to_owned())),
+                },
+            );
+        }
+
+        if let Some(segment_id) = span.segment_id {
+            trace_item.attributes.insert(
+                "sentry.segment_id".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(segment_id.to_owned())),
+                },
+            );
+        }
+
+        if let Some(origin) = span.origin {
+            trace_item.attributes.insert(
+                "sentry.origin".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(origin.to_string())),
+                },
+            );
+        }
+
+        if let Some(kind) = span.kind {
+            trace_item.attributes.insert(
+                "sentry.kind".into(),
+                AnyValue {
+                    value: Some(Value::StringValue(kind.to_owned())),
+                },
+            );
+        }
+
+        self.produce(
+            KafkaTopic::Items,
+            KafkaMessage::Item {
+                headers: BTreeMap::from([
+                    (
+                        "item_type".to_string(),
+                        TraceItemType::Span.as_str_name().to_owned(),
+                    ),
+                    ("project_id".to_string(), scoping.project_id.to_string()),
+                ]),
+                item_type: TraceItemType::Span,
+                message: trace_item,
+            },
+        )?;
+
         Ok(())
     }
 
@@ -1385,7 +1659,7 @@ struct CheckInKafkaMessage {
     retention_days: u16,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 struct SpanLink<'a> {
     pub trace_id: &'a str,
     pub span_id: &'a str,
@@ -1395,16 +1669,16 @@ struct SpanLink<'a> {
     pub attributes: Option<&'a RawValue>,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 struct SpanMeasurement {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     value: Option<f64>,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 struct SpanKafkaMessage<'a> {
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    description: Option<&'a RawValue>,
+    description: Option<String>,
     #[serde(default)]
     duration_ms: u32,
     /// The ID of the transaction event associated to this span, if any.
@@ -1417,8 +1691,8 @@ struct SpanKafkaMessage<'a> {
     #[serde(default)]
     is_remote: bool,
 
-    #[serde(default, skip_serializing_if = "none_or_empty_object")]
-    data: Option<&'a RawValue>,
+    #[serde(default, skip_serializing_if = "none_or_empty_map")]
+    data: Option<BTreeMap<&'a str, JsonValue>>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     kind: Option<&'a str>,
     #[serde(default, skip_serializing_if = "none_or_empty_vec")]
@@ -1450,8 +1724,8 @@ struct SpanKafkaMessage<'a> {
     )]
     sentry_tags: Option<BTreeMap<&'a str, Option<String>>>,
     span_id: &'a str,
-    #[serde(default, skip_serializing_if = "none_or_empty_object")]
-    tags: Option<&'a RawValue>,
+    #[serde(default, skip_serializing_if = "none_or_empty_map")]
+    tags: Option<BTreeMap<&'a str, Option<String>>>,
     trace_id: EventId,
 
     #[serde(default)]
@@ -1513,6 +1787,13 @@ struct LogKafkaMessage<'a> {
     attributes: Option<BTreeMap<&'a str, Option<LogAttribute>>>,
 }
 
+fn none_or_empty_map<K, V>(value: &Option<BTreeMap<K, V>>) -> bool {
+    match value {
+        Some(map) => map.is_empty(),
+        None => true,
+    }
+}
+
 fn none_or_empty_object(value: &Option<&RawValue>) -> bool {
     match value {
         None => true,
@@ -1555,6 +1836,14 @@ enum KafkaMessage<'a> {
     ReplayEvent(ReplayEventKafkaMessage<'a>),
     ReplayRecordingNotChunked(ReplayRecordingNotChunkedKafkaMessage<'a>),
     CheckIn(CheckInKafkaMessage),
+    Item {
+        #[serde(skip)]
+        headers: BTreeMap<String, String>,
+        #[serde(skip)]
+        item_type: TraceItemType,
+        #[serde(skip)]
+        message: TraceItem,
+    },
     Span {
         #[serde(skip)]
         headers: BTreeMap<String, String>,
@@ -1591,6 +1880,7 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::Log { .. } => "log",
             KafkaMessage::Span { .. } => "span",
             KafkaMessage::ProfileChunk(_) => "profile_chunk",
+            KafkaMessage::Item { item_type, .. } => item_type.as_str_name(),
         }
     }
 
@@ -1615,7 +1905,8 @@ impl Message for KafkaMessage<'_> {
             | Self::Log { .. }
             | Self::ReplayRecordingNotChunked(_)
             | Self::ProfileChunk(_)
-            | Self::Metric { .. } => Uuid::nil(),
+            | Self::Metric { .. }
+            | Self::Item { .. } => Uuid::nil(),
         };
 
         if uuid.is_nil() {
@@ -1627,7 +1918,10 @@ impl Message for KafkaMessage<'_> {
 
     fn headers(&self) -> Option<&BTreeMap<String, String>> {
         match &self {
-            KafkaMessage::Metric { headers, .. } => {
+            KafkaMessage::Metric { headers, .. }
+            | KafkaMessage::Span { headers, .. }
+            | KafkaMessage::Item { headers, .. }
+            | KafkaMessage::Log { headers, .. } => {
                 if !headers.is_empty() {
                     return Some(headers);
                 }
@@ -1636,18 +1930,6 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::Profile(profile) => {
                 if !profile.headers.is_empty() {
                     return Some(&profile.headers);
-                }
-                None
-            }
-            KafkaMessage::Log { headers, .. } => {
-                if !headers.is_empty() {
-                    return Some(headers);
-                }
-                None
-            }
-            KafkaMessage::Span { headers, .. } => {
-                if !headers.is_empty() {
-                    return Some(headers);
                 }
                 None
             }
@@ -1667,7 +1949,7 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::Span { message, .. } => serde_json::to_vec(message)
                 .map(Cow::Owned)
                 .map_err(ClientError::InvalidJson),
-            KafkaMessage::Log { message, .. } => {
+            KafkaMessage::Log { message, .. } | KafkaMessage::Item { message, .. } => {
                 let mut payload = Vec::new();
 
                 if message.encode(&mut payload).is_err() {